pyproject.toml
[project]
name = "api-prj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.32.5",
    "pandas",
    "diffusers>=0.36.0",
    "transformers>=5.2.0",
    "torch>=2.10.0",
    "python-dotenv>=1.2.1",
    "flask>=3.1.3",
    "psycopg>=3.3.3",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "curl-cffi>=0.14.0",
    "scikit-learn>=1.8.0",
]
